mutex, doc updates, feature flag
